requirements.txt
streamlit>=1.31.0
langchain>=0.1.14
langchain-community>=0.0.24
openai>=1.10.0
python-dotenv
pandas>=2.2.0
matplotlib>=3.8.0
tqdm
faiss-cpu
PyYAML
